+++ app/Http/Controllers/Staff/AdmissionController.php
@@ -215,7 +215,9 @@ class AdmissionController extends Controller
 
         if(strtolower($status) == 'reverse_admission'){
             $student = Student::where('user_id', $applicantId)->first();
-            unlink($student->admission_letter);
+            if (file_exists($student->admission_letter)) {
+                unlink($student->admission_letter);
+            } 
             $student->forceDelete();
 
             $applicant->status = 'submitted';
